Change JSOutPoint constructor to have js argument be uint64_t
Closes #3192.

+++ src/wallet/wallet.h
@@ -165,7 +165,7 @@ public:
     uint8_t n;
 
     JSOutPoint() { SetNull(); }
-    JSOutPoint(uint256 h, size_t js, uint8_t n) : hash {h}, js {js}, n {n} { }
+    JSOutPoint(uint256 h, uint64_t js, uint8_t n) : hash {h}, js {js}, n {n} { }
 
     ADD_SERIALIZE_METHODS;
